requirements.txt
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
pandas>=2.2.0
click>=8.1.7
pydantic>=2.5.2
openpyxl>=3.1.2
gspread>=5.12.4
colorama>=0.4.6
tabulate>=0.9.0 